Proceed with code extraction
https://trello.com/c/tcIFtuRl
