+++ assets/js/homepage.js
@@ -1,4 +1,6 @@
-import {form, validateForm} from './helper';
+import {form, validateForm, handleModalClose} from './helper';
+
+handleModalClose();
 
 form.addEventListener('submit', (event) => {
     event.preventDefault();
